Add files via upload
Farmer E-commerce Project is developed using PHP, CSS, Bootstrap, and JavaScript. Talking the project, it contains an admin side where he/she can manage all the agricultural and farming records. The admin plays an important role in the management of this system. In this project, the admin has full access to the system. Note down that this system is only available for the admin side for now.